fix(mcp): resolve MCP server import error by consolidating implementations
Fixes broken MCP server that was crashing with "MCPServer export not found"
error when loaded via Claude Desktop. The issue was src/mcp_server.js using
outdated MCP SDK v1.24.0 API (MCPServer class from /server/index.js).

Changes:
- Delete broken src/mcp_server.js with incomplete 5-tool implementation
- Rename src/mcp_server_improved.js → src/mcp_server.js (34-tool impl)
- Update package.json bin entry to point to renamed MCP server file
- Refactor src/index.js to Express-only mode (removed broken MCP startup)
- Update all documentation and test references to use new filename

The improved server uses correct MCP SDK API (McpServer from /server/mcp.js)
and stdio transport, compatible with Claude Desktop and other MCP clients.

Tests: All 327 tests pass, linting passes with no errors.
