+++ src/index.js
@@ -5,7 +5,6 @@ import dotenv from 'dotenv';
 import postRoutes from './routes/postRoutes.js'; // Import post routes
 import imageRoutes from './routes/imageRoutes.js'; // Import image routes
 import tagRoutes from './routes/tagRoutes.js'; // Import tag routes
-import { startMCPServer } from './mcp_server.js'; // Import MCP server start function
 import { createContextLogger } from './utils/logger.js';
 
 // Load environment variables from .env file
@@ -16,7 +15,6 @@ const logger = createContextLogger('main');
 
 const app = express();
 const restApiPort = process.env.PORT || 3000;
-const mcpPort = process.env.MCP_PORT || 3001; // Allow configuring MCP port
 
 // Apply security headers with Helmet (OWASP recommended)
 app.use(
@@ -91,9 +89,8 @@ app.use((err, req, res, _next) => {
   });
 });
 
-// Start both servers
-const startServers = async () => {
-  // Start Express server
+// Start Express server
+const startServer = () => {
   app.listen(restApiPort, () => {
     logger.info('Express REST API server started', {
       port: restApiPort,
@@ -101,16 +98,6 @@ const startServers = async () => {
       type: 'server_start',
     });
   });
-
-  // Start MCP server
-  await startMCPServer(mcpPort);
 };
 
-startServers().catch((error) => {
-  logger.error('Failed to start servers', {
-    error: error.message,
-    stack: error.stack,
-    type: 'server_start_error',
-  });
-  process.exit(1);
-});
+startServer();
